requirements.txt
flask>=2.3.0
Werkzeug >= 3.1.0
flask-cors>=4.0.0
pdfplumber>=0.11.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
groq>=0.8.0
gunicorn>=22.0.0
